requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
opencv-python== 4.11.0.86
face-recognition==1.3.0
dlib==19.24.1
python-dotenv==1.0.0
numpy==1.26.4
requests==2.32.3 